Always use INTERNAL_IP4_ADDRESS as "gateway" on Windows
It appears that Windows's 'netsh' utility doesn't like the VPN interface
having a "gateway" address of 0.0.0.1, which is what vpnc-script-win.js
would attempt to use if INTERNAL_IP4_NETMASK is /0.  See
https://gitlab.com/openconnect/openconnect/-/merge_requests/306#note_745139972.

As noted in the OpenConnect source, "It's a tunnel; having a gateway is
meaningless."

Setting the "gateway" address for Windows to match INTERNAL_IP4_ADDRESS
seems like the simplest way to make 'netsh' configure routingly correctly in
all cases, including when the INTERNAL_IP4_NETMASK is either /0 or /32.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>