remove another unbound cache
Unbound has a separate cache, called "infra cache". This is used for caching
reachability(ping), EDNS and lameness info about forwarders, upstream resolvers
and authoritative servers.

If our forwarders are for some reason cached as unreachable before vpnc configures
them, such as in the case of a frozen connection and then attempting to reconnect,
they'll remain unreachable and all queries attempting to use them will return
SERVFAIL, until cache entries expire.

To fix that, we need to remove our DNS servers from the infra cache on VPN
establishment. That way, unbound will be able to reach the DNS servers and
resolve our queries.

As this is a separate cache, it's not covered by the flush_zone subcommand and
has to be handled differently, through flush_infra.