Detect systemd-resolved in stub mode
This commit fixes systemd-resolved detection for its currently
recommended mode of operation as mentioned in systemd-resolved(8).
In this mode systemd-resolved operates as stub DNS server at 127.0.0.53
and lists itself as nameserver in /etc/resolv.conf for backwards compatibility.
Therefore systemd-resolved should also be considered active
if the hosts line in /etc/nsswitch.conf only lists 'dns',
but /etc/resolv.conf is a symlink to a systemd-resolved managed config.
This mode allows systemd-resolved to be used not only for applications
using GNU NSS, but also others issueing their DNS queries directly.
It is also the current default used in Ubuntu.
The commit also detects the second and third mode mentioned in the
man page in case they are used in other distributions or explicitly
chosen by an administrator.

For further information see:
https://www.freedesktop.org/software/systemd/man/systemd-resolved.service.html#/etc/resolv.conf