vpnc-script-win.js: show the connection banner first, like the POSIX script does
The POSIX version of the script (`vpnc-script`) already shows the connection
banner at the *top* of the `reason=connect` hook.

The Windows script (`vpnc-script-win.js`) has until now shown the connection
banner at the *bottom* of the `reason=connect` hook.

This behaviour is very confusing to Windows users:

1. Unless verbose logging has been enabled, OpenConnect appears to "hang" or
   "stall" after displaying the banner. In reality, it just continues to run
   silently, because it doesn't really need to communicate anything else with
   the user.
2. Windows users aren't used to console applications which remain in the
   foreground.
3. Proprietary clients (particularly Cisco AnyConnect) often make the user
   "accept" the banner before continuing. No one has been able to provide us
   with any evidence that this is anything other than a *purely* client-side
   action; nothing is sent back to the server.

Windows users often get confused by OpenConnect going silent after it prints
the banner, and report that they're unable to "accept" the banner (unlike in
the proprietary clients, or in NetworkManager-openconnect which shows the
banner in a blocking/modal dialog with an "OK" button).

This tweak should "fix" both
https://gitlab.com/openconnect/openconnect/-/issues/1 and
https://gitlab.com/openconnect/openconnect/-/issues/525.  Those aren't
really bugs in OpenConnect or vpnc-script, but just confusion over the
user interface.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>