restore_resolvconf_generic: Fix filtering non-VPN DNS servers on Darwin
restore_resolvconf_generic on Darwin compares DNS servers that scutil
returns for the primary network service with VPN IPv4 DNS servers. If
different, it filters out the latter from the former, by removing a
matching prefix. However this does not work when servers are not listed
in the same order, so it is better to use fgrep.